requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.28.2
botocore==1.31.2
Django==4.2.3
django-storages==1.13.2
jmespath==1.0.1
Pillow==10.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==1.26.16
